Initial splitting. Contains exactly those changes laid out in the issue.
