+++ crates/bevy_hierarchy/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "bevy_transform"
+name = "bevy_hierarchy"
 version = "0.5.0"
 edition = "2018"
 description = "Provides hierarchy and transform functionality for Bevy Engine"
@@ -10,8 +10,9 @@ keywords = ["bevy"]
 
 [dependencies]
 # bevy
-bevy_app = { path = "../bevy_app", version = "0.5.0" }
 bevy_ecs = { path = "../bevy_ecs", version = "0.5.0" }
-bevy_hierarchy = { path = "../bevy_hierarchy", version = "0.5.0" }
-bevy_math = { path = "../bevy_math", version = "0.5.0" }
 bevy_reflect = { path = "../bevy_reflect", version = "0.5.0", features = ["bevy"] }
+bevy_utils = { path = "../bevy_utils", version = "0.5.0" }
+
+# other
+smallvec = { version = "1.6", features = ["serde", "union", "const_generics"] }
