+++ crates/bevy_hierarchy/README.md
@@ -1,0 +1,3 @@
+# Bevy Hierarchy
+
+Provides the ability to work with a parent-child hierarchy of entities.
