+++ crates/bevy_hierarchy/src/hierarchy/hierarchy_maintenance_system.rs
@@ -71,101 +71,3 @@ pub fn parent_update_system(
         commands.entity(*e).insert(Children::with(v));
     });
 }
-#[cfg(test)]
-mod test {
-    use bevy_ecs::{
-        schedule::{Schedule, Stage, SystemStage},
-        system::CommandQueue,
-        world::World,
-    };
-
-    use super::*;
-    use crate::{hierarchy::BuildChildren, transform_propagate_system::transform_propagate_system};
-
-    #[test]
-    fn correct_children() {
-        let mut world = World::default();
-
-        let mut update_stage = SystemStage::parallel();
-        update_stage.add_system(parent_update_system);
-        update_stage.add_system(transform_propagate_system);
-
-        let mut schedule = Schedule::default();
-        schedule.add_stage("update", update_stage);
-
-        // Add parent entities
-        let mut command_queue = CommandQueue::default();
-        let mut commands = Commands::new(&mut command_queue, &world);
-        let mut children = Vec::new();
-        let parent = commands
-            .spawn()
-            .insert(Transform::from_xyz(1.0, 0.0, 0.0))
-            .id();
-        commands.entity(parent).with_children(|parent| {
-            children.push(
-                parent
-                    .spawn()
-                    .insert(Transform::from_xyz(0.0, 2.0, 0.0))
-                    .id(),
-            );
-            children.push(
-                parent
-                    .spawn()
-                    .insert(Transform::from_xyz(0.0, 3.0, 0.0))
-                    .id(),
-            );
-        });
-        command_queue.apply(&mut world);
-        schedule.run(&mut world);
-
-        assert_eq!(
-            world
-                .get::<Children>(parent)
-                .unwrap()
-                .0
-                .iter()
-                .cloned()
-                .collect::<Vec<_>>(),
-            children,
-        );
-
-        // Parent `e1` to `e2`.
-        (*world.get_mut::<Parent>(children[0]).unwrap()).0 = children[1];
-
-        schedule.run(&mut world);
-
-        assert_eq!(
-            world
-                .get::<Children>(parent)
-                .unwrap()
-                .iter()
-                .cloned()
-                .collect::<Vec<_>>(),
-            vec![children[1]]
-        );
-
-        assert_eq!(
-            world
-                .get::<Children>(children[1])
-                .unwrap()
-                .iter()
-                .cloned()
-                .collect::<Vec<_>>(),
-            vec![children[0]]
-        );
-
-        assert!(world.despawn(children[0]));
-
-        schedule.run(&mut world);
-
-        assert_eq!(
-            world
-                .get::<Children>(parent)
-                .unwrap()
-                .iter()
-                .cloned()
-                .collect::<Vec<_>>(),
-            vec![children[1]]
-        );
-    }
-}
