+++ crates/bevy_hierarchy/src/lib.rs
@@ -1,0 +1,7 @@
+pub mod components;
+pub mod hierarchy;
+
+pub mod prelude {
+    #[doc(hidden)]
+    pub use crate::{components::*, hierarchy::*};
+}
