+++ crates/bevy_internal/src/lib.rs
@@ -29,6 +29,11 @@ pub mod ecs {
     pub use bevy_ecs::*;
 }
 
+pub mod hierarchy {
+    //! Parent-child hierarchy.
+    pub use bevy_hierarchy::*;
+}
+
 pub mod input {
     //! Resources and events for inputs, e.g. mouse/keyboard, touch, gamepads, etc.
     pub use bevy_input::*;
