+++ crates/bevy_transform/src/transform_propagate_system.rs
@@ -1,9 +1,10 @@
-use crate::components::{Children, GlobalTransform, Parent, Transform};
+use crate::components::{GlobalTransform, Transform};
 use bevy_ecs::{
     entity::Entity,
     query::{Changed, With, Without},
     system::Query,
 };
+use bevy_hierarchy::prelude::{Children, Parent};
 
 /// Update [`GlobalTransform`] component of entities based on entity hierarchy and
 /// [`Transform`] component.
@@ -24,7 +25,7 @@ pub fn transform_propagate_system(
         }
 
         if let Some(children) = children {
-            for child in children.0.iter() {
+            for child in children.iter() {
                 propagate_recursive(
                     &global_transform,
                     &changed_transform_query,
@@ -60,7 +61,7 @@ fn propagate_recursive(
     };
 
     if let Ok(Some(children)) = children_query.get(entity) {
-        for child in children.0.iter() {
+        for child in children.iter() {
             propagate_recursive(
                 &global_matrix,
                 changed_transform_query,
@@ -82,7 +83,7 @@ mod test {
     };
 
     use super::*;
-    use crate::hierarchy::{parent_update_system, BuildChildren, BuildWorldChildren};
+    use bevy_hierarchy::prelude::{parent_update_system, BuildChildren, BuildWorldChildren};
 
     #[test]
     fn did_propagate() {
@@ -188,6 +189,107 @@ mod test {
         assert_eq!(
             *world.get::<GlobalTransform>(children[1]).unwrap(),
             GlobalTransform::from_xyz(1.0, 0.0, 0.0) * Transform::from_xyz(0.0, 0.0, 3.0)
+        );
+    }
+}
+
+#[cfg(test)]
+mod test2 {
+    use super::Transform;
+    use bevy_ecs::{
+        schedule::{Schedule, Stage, SystemStage},
+        system::CommandQueue,
+        world::World,
+    };
+    use bevy_hierarchy::prelude::{parent_update_system, BuildChildren, Children, Parent};
+
+    use bevy_ecs::system::Commands;
+
+    use crate::transform_propagate_system::transform_propagate_system;
+
+    #[test]
+    fn correct_children() {
+        let mut world = World::default();
+
+        let mut update_stage = SystemStage::parallel();
+        update_stage.add_system(parent_update_system);
+        update_stage.add_system(transform_propagate_system);
+
+        let mut schedule = Schedule::default();
+        schedule.add_stage("update", update_stage);
+
+        // Add parent entities
+        let mut command_queue = CommandQueue::default();
+        let mut commands = Commands::new(&mut command_queue, &world);
+        let mut children = Vec::new();
+        let parent = commands
+            .spawn()
+            .insert(Transform::from_xyz(1.0, 0.0, 0.0))
+            .id();
+        commands.entity(parent).with_children(|parent| {
+            children.push(
+                parent
+                    .spawn()
+                    .insert(Transform::from_xyz(0.0, 2.0, 0.0))
+                    .id(),
+            );
+            children.push(
+                parent
+                    .spawn()
+                    .insert(Transform::from_xyz(0.0, 3.0, 0.0))
+                    .id(),
+            );
+        });
+        command_queue.apply(&mut world);
+        schedule.run(&mut world);
+
+        assert_eq!(
+            world
+                .get::<Children>(parent)
+                .unwrap()
+                .iter()
+                .cloned()
+                .collect::<Vec<_>>(),
+            children,
+        );
+
+        // Parent `e1` to `e2`.
+        (*world.get_mut::<Parent>(children[0]).unwrap()).0 = children[1];
+
+        schedule.run(&mut world);
+
+        assert_eq!(
+            world
+                .get::<Children>(parent)
+                .unwrap()
+                .iter()
+                .cloned()
+                .collect::<Vec<_>>(),
+            vec![children[1]]
+        );
+
+        assert_eq!(
+            world
+                .get::<Children>(children[1])
+                .unwrap()
+                .iter()
+                .cloned()
+                .collect::<Vec<_>>(),
+            vec![children[0]]
+        );
+
+        assert!(world.despawn(children[0]));
+
+        schedule.run(&mut world);
+
+        assert_eq!(
+            world
+                .get::<Children>(parent)
+                .unwrap()
+                .iter()
+                .cloned()
+                .collect::<Vec<_>>(),
+            vec![children[1]]
         );
     }
 }
