+++ crates/bevy_ui/Cargo.toml
@@ -15,6 +15,7 @@ bevy_asset = { path = "../bevy_asset", version = "0.5.0" }
 bevy_core = { path = "../bevy_core", version = "0.5.0" }
 bevy_derive = { path = "../bevy_derive", version = "0.5.0" }
 bevy_ecs = { path = "../bevy_ecs", version = "0.5.0" }
+bevy_hierarchy = { path = "../bevy_hierarchy", version = "0.5.0" }
 bevy_input = { path = "../bevy_input", version = "0.5.0" }
 bevy_log = { path = "../bevy_log", version = "0.5.0" }
 bevy_math = { path = "../bevy_math", version = "0.5.0" }
