+++ crates/bevy_ui/src/flex/mod.rs
@@ -7,9 +7,10 @@ use bevy_ecs::{
     query::{Changed, FilterFetch, With, Without, WorldQuery},
     system::{Query, Res, ResMut},
 };
+use bevy_hierarchy::prelude::{Children, Parent};
 use bevy_log::warn;
 use bevy_math::Vec2;
-use bevy_transform::prelude::{Children, Parent, Transform};
+use bevy_transform::prelude::Transform;
 use bevy_utils::HashMap;
 use bevy_window::{Window, WindowId, WindowScaleFactorChanged, Windows};
 use std::fmt;
